requirements.txt
coverage==7.6.4
python-dotenv==1.0.1
et_xmlfile~=2.0.0
iniconfig~=2.0.0
numpy~=2.1.3
openpyxl~=3.1.5
packaging~=24.2
pandas~=2.2.3
pluggy~=1.5.0
pytest~=8.3.3
python-dateutil~=2.9.0.post0
pytz~=2024.2
PyYAML~=6.0.2
six~=1.16.0
typing_extensions~=4.12.2
tzdata~=2024.2
u~=0.1
websockets~=14.0
wheel==0.45.0